fix(search): preserve u64 lane encoding in StarlarkLane.encoded getter
`StarlarkLane.encoded` is the third instance of the same i32-truncation
bug class as the two we fixed earlier in
`crates/bloqade-lanes-search/src/move_policy_dsl/{actions.rs,kernel.rs}`.
The attribute getter cast `self.0.encode_u64() as i32`, with an inline
comment asserting "Gemini lane IDs fit in i32 in practice." They don't
on the physical Gemini arch — `LaneAddr::encode_u64` packs zone/bus/site
bits into the high half of a u64, producing values like
`2305843009213693952 ≈ 2^61`. Casting to i32 truncates the lower 32
bits — all zero for high-bit-only encodings — yielding `lane.encoded =
0`. Every `insert_child(node, [lane.encoded])` then attempts to execute
"lane 0", which `MoveAction::InsertChild` rejects with
`aod_invalid: lane 0 has no qubit at src`, and the policy makes no
progress.

This bug fundamentally blocks DSL search policies that use `lane.encoded`
directly (rather than going through `pack_aod_rectangles → move_set.encoded`).
Verified end-to-end by a probe DFS that bypasses `pack_aod_rectangles`:

  Before: 0 nodes expanded on ghz_4; `lanes: [0]` in every InsertChild.
  After:  74 nodes expanded on ghz_4; real lane encodings like
          2305843009213693952, 65536, 9223372036854841344.

Discovered while debugging why every search-based DSL reference policy
(dfs.star, bfs.star, ids.star) halts at ~5 nodes on physical kernels.
Note: a separate bug in `pack_aod_rectangles::is_valid_rect` —
`collision_srcs` rejects any source whose bus has any outgoing lane to
an occupied destination, even for 1x1 rectangles — still blocks the
pack_aod_rectangles-based policies. Tracked separately.

+++ crates/bloqade-lanes-dsl-core/src/primitives/types.rs
@@ -90,8 +90,15 @@ impl fmt::Display for StarlarkLane {
 impl<'v> StarlarkValue<'v> for StarlarkLane {
     fn get_attr(&self, attr: &str, heap: &'v Heap) -> Option<Value<'v>> {
         match attr {
-            // encode_u64 returns u64; Gemini lane IDs fit in i32 in practice.
-            "encoded" => Some(heap.alloc(self.0.encode_u64() as i32)),
+            // Physical-Gemini lane encodings routinely exceed `i32::MAX` (the
+            // `LaneAddr::encode_u64` packs zone/bus/site bits into the high
+            // half of a `u64`). Casting to `i32` here silently wraps the
+            // value — typically to `0` for the high-bit-only encodings —
+            // and the resulting `lanes: [0]` makes every `insert_child`
+            // call fail with `aod_invalid: lane 0 has no qubit at src`,
+            // silently corrupting every DSL search policy that reads
+            // `lane.encoded` to forward to `insert_child`.
+            "encoded" => Some(heap.alloc(self.0.encode_u64() as i64)),
             // direction and move_type are #[repr(u8)] enums; cast to i32.
             "direction" => Some(heap.alloc(self.0.direction as i32)),
             "move_type" => Some(heap.alloc(self.0.move_type as i32)),
